pyproject.toml
[project]
name = "langgraph-first-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "ipykernel>=7.2.0",
    "langchain[anthropic,chatopenai]>=1.2.15",
    "langgraph>=1.1.8",
    "python-dotenv>=1.2.2",
]
